v1.4 crossref: add hugo-md support
