+++ tests/docs/smoke-all/crossrefs/float/hugo/hugo-float-5.qmd
@@ -1,0 +1,18 @@
+---
+title: Crossref Test
+format: hugo-md
+_quarto:
+  tests:
+    hugo-md:
+      ensureFileRegexMatches:
+        - 
+          - "<img src=\"img/thinker.jpg\" id=\"fig-elephant\" alt=\"Figure 1: Elephant\" />"
+        - 
+          - "<a href=\"#fig-elephant1\" class=\"quarto-xref\">"
+---
+
+## Unresolved Crossref Figure
+
+![Elephant](img/thinker.jpg){#fig-elephant}
+
+See @fig-elephant1 for examples.
