+++ tests/docs/smoke-all/crossrefs/float/hugo/hugo-float-numbering-1.qmd
@@ -1,0 +1,46 @@
+---
+title: "My Document"
+format: hugo-md
+crossref:
+  fig-labels: alpha x # (default is arabic)
+  tbl-labels: alpha A # (default is arabic)
+  subref-labels: roman i # (default is alpha a)
+_quarto:
+  tests:
+    hugo-md:
+      ensureFileRegexMatches:
+        - 
+          - "Figure&nbsp;x: Elephant"
+          - "Table&nbsp;A: My Caption"
+          - "Figure&nbsp;y: Famous Elephants"
+          - "\\(i\\) Surus"
+          - "\\(ii\\) Abbas"
+---
+
+# Introduction
+
+![Elephant](img/painter.jpg){#fig-elephant}
+
+See @fig-elephant for an illustration.
+
+| Col1 | Col2 | Col3 |
+| ---- | ---- | ---- |
+| A    | B    | C    |
+| E    | F    | G    |
+| A    | G    | G    |
+
+: My Caption {#tbl-letters}
+
+See @tbl-letters.
+
+## Simple Sub Figure
+
+::: {#fig-elephants layout-ncol="2"}
+![Surus](img/surus.jpg){#fig-surus}
+
+![Abbas](img/abbas.jpg){#fig-abbas}
+
+Famous Elephants
+:::
+
+See @fig-elephants for examples. In particular, @fig-abbas.
