+++ tests/docs/smoke-all/crossrefs/float/hugo/hugo-float-options-1.qmd
@@ -1,0 +1,36 @@
+---
+title: "My Document"
+format: hugo-md
+crossref:
+  fig-prefix: F.   # (default is "fig.")
+  tbl-prefix: T.    # (default is "tbl.")
+  title-delim: —     # (default is ":")
+  ref-hyperlink: false # (default is true)
+_quarto:
+  tests:
+    hugo-md:
+      ensureFileRegexMatches:
+        - 
+          - "Figure&nbsp;1— Elephant"
+          - "See T.&nbsp;1."
+          - "See F.&nbsp;1."
+        -
+          - "<a href=\"#tbl-letters\" class=\"quarto-xref\">" # ref-hyperlink: false
+---
+
+# Introduction
+
+![Elephant](img/painter.jpg){#fig-elephant}
+
+See @fig-elephant for an illustration.
+
+
+| Col1 | Col2 | Col3 |
+|------|------|------|
+| A    | B    | C    |
+| E    | F    | G    |
+| A    | G    | G    |
+
+: My Caption {#tbl-letters}
+
+See @tbl-letters.
