+++ tests/docs/smoke-all/crossrefs/float/hugo/hugo-jupyter-2.qmd
@@ -1,0 +1,36 @@
+---
+title: Python Subfig Test
+format: hugo-md
+_quarto:
+  tests:
+    hugo-md:
+      ensureFileRegexMatches:
+        - 
+          - "id=\"fig-plots-1\""
+          - "alt=\"\\(a\\) Plot 1\""
+          - "id=\"fig-plots-2\""
+          - "alt=\"\\(b\\) Plot 2\""
+          - "<a href=\"#fig-plots\" class=\"quarto-xref\">"
+          - "<a href=\"#fig-plots-2\" class=\"quarto-xref\">"
+        - []
+---
+
+## Python Crossref Figure
+
+```{python}
+#| label: fig-plots
+#| fig-cap: "Plots" 
+#| fig-subcap:
+#|   - "Plot 1"
+#|   - "Plot 2" 
+#| layout-ncol: 2
+
+import matplotlib.pyplot as plt
+plt.plot([1,23,2,4])
+plt.show()
+
+plt.plot([8,65,23,90])
+plt.show()
+```
+
+See @fig-plots for examples. In particular, @fig-plots-2.
