+++ tests/docs/smoke-all/crossrefs/float/hugo/hugo-knitr-1.qmd
@@ -1,0 +1,23 @@
+---
+title: Knitr Crossref Test
+format: hugo-md
+_quarto:
+  tests:
+    hugo-md:
+      ensureFileRegexMatches:
+        - 
+          - "<a href=\"#fig-plot\" class=\"quarto-xref\">"
+          - "id=\"fig-plot\""
+          - "alt=\"Figure 1: Plot\""
+---
+
+## Knitr Crossref Figure
+
+```{r}
+#| label: fig-plot
+#| fig-cap: "Plot"
+
+plot(cars)
+```
+
+For example, see @fig-plot.
