+++ tests/docs/smoke-all/crossrefs/float/hugo/hugo-ojs-lst-cap.qmd
@@ -1,0 +1,20 @@
+---
+title: listing within R chunks
+format: hugo-md
+# keep-md: true
+# FIXME cscheid needs to fix this on compile.ts
+# _quarto:
+#   tests:
+#     html:
+#       ensureHtmlElements:
+#         - ['a.quarto-xref[href="#lst-code-source"]', 'div#lst-code-source.listing']
+#         - []
+---
+
+```{ojs}
+//| lst-label: lst-code-source
+//| lst-cap: Some code source
+1 + 1
+```
+
+See @lst-code-source.
