+++ tests/docs/smoke-all/crossrefs/float/hugo/hugo-subfloat-1.qmd
@@ -1,0 +1,32 @@
+---
+title: Crossref Test
+format: hugo-md
+_quarto:
+  tests:
+    hugo-md:
+      ensureFileRegexMatches:
+        -
+          - "\\(a\\) Surus"
+          - "\\(b\\) Abbas"
+          - "<img src=\"img/surus.jpg\" id=\"fig-surus\" alt=\"\\(a\\) Surus\" />"
+          - "<img src=\"img/abbas.jpg\" id=\"fig-abbas\" alt=\"\\(b\\) Abbas\" />"
+#       ensureHtmlElements:
+#         - 
+#           - "div#fig-elephants figure.quarto-float-fig > figcaption.quarto-float-caption"
+#           - "div#fig-surus figure.quarto-subfloat-fig figcaption.quarto-subfloat-caption"
+#           - "div#fig-abbas figure.quarto-subfloat-fig figcaption.quarto-subfloat-caption"
+#         - []
+---
+
+## Simple Sub Figure
+
+::: {#fig-elephants layout-ncol=2}
+
+![Surus](img/surus.jpg){#fig-surus}
+
+![Abbas](img/abbas.jpg){#fig-abbas}
+
+Famous Elephants
+:::
+
+See @fig-elephants for examples. In particular, @fig-abbas and @fig-surus.
