+++ tests/docs/smoke-all/crossrefs/float/hugo/rawtablecaption.qmd
@@ -1,0 +1,32 @@
+---
+title: table caption test
+format: hugo-md
+_quarto:
+  tests:
+    hugo-md:
+      ensureFileRegexMatches:
+        - 
+          - "<div id=\"tbl-1\">"
+          - "Table 1: This is a caption."
+          - "<a href=\"#tbl-1\" class=\"quarto-xref\">"
+        - []
+---
+
+::: {#tbl-1}
+
+```{=html}
+<table>
+  <caption>This is a caption.</caption>
+  <thead>
+  <tr><th>col 1</th><th>col 2</th></tr>
+  </thead>
+  <tbody>
+  <tr><td>1</td><td>2</td></tr>
+  <tr><td>3</td><td>4</td></tr>
+  </tbody>
+</table>
+```
+
+:::
+
+See @tbl-1.
